more refactoring, probably final cleanup of Compare.js for a while
