+++ Frontend/website/src/components/pages/Compare.js
@@ -4,68 +4,15 @@
 
 import "./ComparePage.css";
 import React from "react";
-import {
-	CompareContextProvider,
-	useCompareContext,
-} from "./compare_components/CompareContext";
-import FileUpload from "./compare_components/FileUpload";
-import URLBox from "./compare_components/URLBox";
+import { CompareContextProvider } from "./compare_components/CompareContext";
 import ContrastBar from "./compare_components/ContrastBar";
-import {
-	ArticleSentimentBar1,
-	ArticleSentimentBar2,
-} from "./compare_components/ArticleSentimentBar";
 import Legend from "./compare_components/Legend";
 import Summary from "./compare_components/Summary";
 import CompareButton from "./compare_components/CompareButton";
 import AspectBox from "./compare_components/AspectBox";
+import UploadBox from "./compare_components/UploadBox";
+import { ArticlePane1, ArticlePane2 } from "./compare_components/ArticlePane";
 
-function UploadBox() {
-	const { file1Label, file2Label } = useCompareContext();
-
-	return (
-		<div className='upload'>
-			<ul>
-				<li>
-					<div>
-						<URLBox id='urlBox1' placeholder={file1Label} />
-					</div>
-				</li>
-				<li>
-					<FileUpload id='file1' />
-				</li>
-			</ul>
-			<ul className='upload2'>
-				<li>
-					<div>
-						<URLBox id='urlBox2' placeholder={file2Label} />
-					</div>
-				</li>
-				<li>
-					<FileUpload id='file2' />
-				</li>
-			</ul>
-		</div>
-	);
-}
-
-
-
-function ArticlePane1() {
-	return (
-		<div className='articlePane'>
-			<ArticleSentimentBar1 />
-		</div>
-	);
-}
-
-function ArticlePane2() {
-	return (
-		<div className='articlePane'>
-			<ArticleSentimentBar2 />
-		</div>
-	);
-}
 
 function Compare() {
 	return (
